requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
mysql-connector-python==8.2.0
selenium==4.16.0
python-dotenv==1.0.0 